# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  templatev25_internal_http_dto.ChatItemCreateDto:
    properties:
      answer:
        type: string
      key:
        type: string
    required:
    - answer
    - key
    type: object
  templatev25_internal_http_dto.ChatItemKeyDto:
    properties:
      key:
        type: string
    required:
    - key
    type: object
  templatev25_internal_http_dto.ChatItemUpdateDto:
    properties:
      answer:
        type: string
      key:
        type: string
    required:
    - answer
    - key
    type: object
...
info:
  contact: {}
  description: Fiber + Gorm + Auth + RBAC demo
  title: Template Backend API
  version: "1.0"
paths:
...
  /chat:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List chat items
      tags:
      - chat
    post:
      consumes:
      - application/json
      parameters:
      - description: Chat item data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/templatev25_internal_http_dto.ChatItemCreateDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create chat item
      tags:
      - chat
  /chat/{id}:
    delete:
      parameters:
      - description: Chat item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete chat item
      tags:
      - chat
    put:
      consumes:
      - application/json
      parameters:
      - description: Chat item ID
        in: path
        name: id
        required: true
        type: integer
      - description: Chat item data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/templatev25_internal_http_dto.ChatItemUpdateDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update chat item
      tags:
      - chat
  /chat/key:
    post:
      consumes:
      - application/json
      parameters:
      - description: Key
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/templatev25_internal_http_dto.ChatItemKeyDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get chat item by key
      tags:
      - chat
  /client:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List OAuth clients
      tags:
      - client
  /client/scope:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Scope data
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete OAuth scope
      tags:
      - client
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List OAuth scopes
      tags:
      - client
    post:
      consumes:
      - application/json
      parameters:
      - description: Scope data
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create OAuth scope
      tags:
      - client
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"